# cmd/server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.OperationWithNameAndType:
    properties:
      name:
        type: string
      type:
        type: string
    required:
    - name
    - type
    type: object
...
  request.FileBatchDelete:
    properties:
      isDir:
        type: boolean
      paths:
        items:
          type: string
        type: array
    required:
    - paths
    type: object
...
  request.FileCreate:
    properties:
      content:
        type: string
      isDir:
        type: boolean
      isLink:
        type: boolean
      isSymlink:
        type: boolean
      linkPath:
        type: string
      mode:
        type: integer
      path:
        type: string
      sub:
        type: boolean
    required:
    - mode
    - path
    type: object
  request.FileDeCompress:
    properties:
      dst:
        type: string
      path:
        type: string
      type:
        type: string
    required:
    - dst
    - path
    - type
    type: object
...
  request.FileDownload:
    properties:
      compress:
        type: boolean
      name:
        type: string
      paths:
        items:
          type: string
        type: array
      type:
        type: string
    required:
    - compress
    - name
    - paths
    - type
    type: object
...
  request.FilePathCheck:
    properties:
      path:
        type: string
    required:
    - path
    type: object
...
host: localhost
info:
  contact: {}
  description: 开源Linux面板
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 1Panel
  version: "1.0"
paths:
...
  /containers/daemonjson/file:
    get:
      description: 获取 docker 配置信息(表单)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Load docker daemon.json
      tags:
      - Container Docker
...
  /databases/load/file:
    post:
      consumes:
      - application/json
      description: 获取数据库文件
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.OperationWithNameAndType'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Load Database file
      tags:
      - Database
...
  /files:
    post:
      consumes:
      - application/json
      description: 创建文件/文件夹
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.FileCreate'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Create file
      tags:
      - File
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - path
        formatEN: Create dir or file [path]
        formatZH: 创建文件/文件夹 [path]
        paramKeys: []
  /files/batch/del:
    post:
      consumes:
      - application/json
      description: 批量删除文件/文件夹
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.FileBatchDelete'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Batch delete file
      tags:
      - File
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - paths
        formatEN: Batch delete dir or file [paths]
        formatZH: 批量删除文件/文件夹 [paths]
        paramKeys: []
  /files/check:
    post:
      consumes:
      - application/json
      description: 检测文件是否存在
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.FilePathCheck'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Check file exist
      tags:
      - File
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - path
        formatEN: Check whether file [path] exists
        formatZH: 检测文件 [path] 是否存在
        paramKeys: []
  /files/chunkdownload:
    post:
      consumes:
      - application/json
      description: 分片下载下载文件
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.FileDownload'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Chunk Download file
      tags:
      - File
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - name
        formatEN: Download file [name]
        formatZH: 下载文件 [name]
        paramKeys: []
  /files/chunkupload:
    post:
      description: 分片上传文件
      parameters:
      - description: request
        in: formData
        name: file
        required: true
        type: file
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: ChunkUpload file
      tags:
      - File
  /files/compress:
    post:
      consumes:
      - application/json
      description: 压缩文件
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.FileCompress'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Compress file
      tags:
      - File
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - name
        formatEN: Compress file [name]
        formatZH: 压缩文件 [name]
        paramKeys: []
  /files/content:
    post:
      consumes:
      - application/json
      description: 获取文件内容
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.FileOption'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.FileInfo'
      security:
      - ApiKeyAuth: []
      summary: Load file content
      tags:
      - File
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - path
        formatEN: Load file content [path]
        formatZH: 获取文件内容 [path]
        paramKeys: []
  /files/decompress:
    post:
      consumes:
      - application/json
      description: 解压文件
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.FileDeCompress'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Decompress file
      tags:
      - File
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - path
        formatEN: Decompress file [path]
        formatZH: 解压 [path]
        paramKeys: []
  /files/del:
    post:
      consumes:
      - application/json
      description: 删除文件/文件夹
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.FileDelete'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Delete file
      tags:
      - File
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - path
        formatEN: Delete dir or file [path]
        formatZH: 删除文件/文件夹 [path]
        paramKeys: []
  /files/download:
    post:
      consumes:
      - application/json
      description: 下载文件
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.FileDownload'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Download file
      tags:
      - File
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - name
        formatEN: Download file [name]
        formatZH: 下载文件 [name]
        paramKeys: []
  /files/mode:
    post:
      consumes:
      - application/json
      description: 修改文件权限
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.FileCreate'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Change file mode
      tags:
      - File
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - path
        - mode
        formatEN: Change mode [paths] => [mode]
        formatZH: 修改权限 [paths] => [mode]
        paramKeys: []
  /files/move:
    post:
      consumes:
      - application/json
      description: 移动文件
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.FileMove'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Move file
      tags:
      - File
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - oldPaths
        - newPath
        formatEN: Move [oldPaths] => [newPath]
        formatZH: 移动文件 [oldPaths] => [newPath]
        paramKeys: []
  /files/owner:
    post:
      consumes:
      - application/json
      description: 修改文件用户/组
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.FileRoleUpdate'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Change file owner
      tags:
      - File
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - path
        - user
        - group
        formatEN: Change owner [paths] => [user]/[group]
        formatZH: 修改用户/组 [paths] => [user]/[group]
        paramKeys: []
  /files/rename:
    post:
      consumes:
      - application/json
      description: 修改文件名称
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.FileRename'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Change file name
      tags:
      - File
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - oldName
        - newName
        formatEN: Rename [oldName] => [newName]
        formatZH: 重命名 [oldName] => [newName]
        paramKeys: []
  /files/save:
    post:
      consumes:
      - application/json
      description: 更新文件内容
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.FileEdit'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update file content
      tags:
      - File
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - path
        formatEN: Update file content [path]
        formatZH: 更新文件内容 [path]
        paramKeys: []
  /files/search:
    post:
      consumes:
      - application/json
      description: 获取文件列表
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.FileOption'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.FileInfo'
      security:
      - ApiKeyAuth: []
      summary: List files
      tags:
      - File
  /files/size:
    post:
      consumes:
      - application/json
      description: 获取文件夹大小
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.DirSizeReq'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Load file size
      tags:
      - File
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - path
        formatEN: Load file size [path]
        formatZH: 获取文件夹大小 [path]
        paramKeys: []
...
  /logs/system/files:
    get:
      description: 获取系统日志文件列表
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Load system log files
      tags:
      - Logs
...
swagger: "2.0"